update lab 7

+++ Lab7/Lab7_ex3/src/main/java/com/example/lab7_ex3/Service/StudentService.java
@@ -1,6 +1,8 @@
 package com.example.lab7_ex3.Service;
 
 import com.example.lab7_ex3.Model.Student;
+import org.springframework.stereotype.Controller;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
@@ -8,5 +10,5 @@ public interface StudentService {
     void add(Student student);
     List<Student> getAll();
 
-    void  delete(long id);
+    void  delete(int id);
 }
